Refactor out FSM generation into build functions
Although this introduces some helper functions, it does not reduce readability and has several key advantages. Firstly, it makes it easier to see the underlying grammar at a glance from the parser combinator functions. More importantly, it reduces the verbosity of the test code by 80-90%, since we do not need to copy-paste the logic for how to generate an Or every time we test nested infix operators for example.